Extract headless DocViewModel base; rebase desktop DocsPanelModel onto it
Creates `core/docs/DocViewModel` as a shared, platform-neutral HoistModel base that owns:
active doc state (activeDoc, content, activeSection, pendingScrollSection), loadContentTask,
the five shared computeds (activeSource/activeCategory/activeCategorySiblings/activeSourceCategories/sections),
navigateToDoc/navigateToCategory/setActiveSection/clearPendingScrollSection/resolveContentLink,
bidirectional route<->doc sync (updateDocFromRoute/updateRouteFromDoc), and loadContentAsync.
An onDocActivated(entry) hook lets subclasses react to doc changes.

DocsPanelModel now extends DocViewModel and retains only desktop-specific members: gridModel,
dockContainerModel, navPanelModel, search state and methods, feedback methods, buildTreeData/loadNav,
createGridModel, getCategoryIcon/getSourceIcon, and onSelectionChange. It overrides onDocActivated
to sync grid selection and exit search mode. Both classes call makeObservable(this) in their own
constructors for their respective observables.
