+++ client-app/src/desktop/tabs/docs/DocsPanelModel.ts
@@ -1,28 +1,21 @@
 import {GridModel} from '@xh/hoist/cmp/grid';
-import {Content, HoistModel, managed, TaskObserver, XH} from '@xh/hoist/core';
+import {Content, managed, XH} from '@xh/hoist/core';
 import {DockContainerModel} from '@xh/hoist/desktop/cmp/dock';
 import {PanelModel} from '@xh/hoist/desktop/cmp/panel';
 import {Icon} from '@xh/hoist/icon';
 import {action, bindable, computed, makeObservable, observable, runInAction} from '@xh/hoist/mobx';
-import {DocCategory, DocEntry, DocExampleLink, getDocExamples} from './docRegistry';
-import {DocSearchResult, DocService} from '../../../core/svc/DocService';
-
-export interface DocSection {
-    id: string;
-    title: string;
-}
+import {DocViewModel} from '../../../core/docs/DocViewModel';
+import {DocEntry, DocExampleLink, getDocExamples} from './docRegistry';
+import {DocSearchResult} from '../../../core/svc/DocService';
 
 /**
  * Primary model for the Docs viewer tab.
  *
- * Manages a tree-based navigation grid of hoist-react and hoist-core documentation,
- * full-text search mode with ranked results, and loading of markdown content
- * for the selected doc. Active doc selection is synced with the URL via a
- * route parameter (e.g. /app/docs/core).
+ * Extends DocViewModel for shared viewer state and extends it with desktop-specific chrome:
+ * a tree-based navigation grid, full-text search mode with ranked results, and a feedback dock.
+ * Active doc selection is synced with the URL via a route parameter (e.g. /app/docs/core).
  */
-export class DocsPanelModel extends HoistModel {
-    private readonly BASE_ROUTE = 'default.docs';
-
+export class DocsPanelModel extends DocViewModel {
     @managed
     gridModel: GridModel;
 
@@ -52,31 +45,8 @@ export class DocsPanelModel extends HoistModel {
     @observable
     selectedSearchIdx: number = -1;
 
-    @observable.ref
-    activeDoc: DocEntry = null;
-
-    @observable.ref
-    content: string = null;
-
-    @managed
-    loadContentTask: TaskObserver = TaskObserver.trackLast();
-
-    @observable
-    activeSection: string = null;
-
-    /**
-     * Section slug a deep-link has requested we scroll to once content renders. Consumed and
-     * cleared by the view; not persisted to the URL (sections are scroll-to-on-arrival targets).
-     */
-    @observable
-    pendingScrollSection: string = null;
-
     @bindable
     feedbackMessage: string = '';
-
-    private get docService(): DocService {
-        return DocService.instance;
-    }
 
     constructor() {
         super();
@@ -97,10 +67,6 @@ export class DocsPanelModel extends HoistModel {
             {
                 track: () => this.gridModel.selectedRecord,
                 run: rec => this.onSelectionChange(rec)
-            },
-            {
-                track: () => XH.routerState.params,
-                run: () => this.updateDocFromRoute()
             }
         );
     }
@@ -109,90 +75,17 @@ export class DocsPanelModel extends HoistModel {
         this.loadNav();
     }
 
-    /** The source info for the currently active doc. */
-    @computed
-    get activeSource(): string | null {
-        return this.activeDoc?.source ?? null;
-    }
-
-    /** The category of the currently active doc. */
-    @computed
-    get activeCategory(): DocCategory | null {
-        if (!this.activeDoc) return null;
-        const cats = this.docService.getCategories(this.activeDoc.source);
-        return cats.find(c => c.id === this.activeDoc.category) ?? null;
-    }
-
-    /** Sibling docs in the same source + category as the active doc. */
-    @computed
-    get activeCategorySiblings(): DocEntry[] {
-        if (!this.activeDoc) return [];
-        return this.docService.getDocsByCategory(this.activeDoc.source, this.activeDoc.category);
-    }
-
-    /** Categories for the active doc's source, filtered to those with at least one doc. */
-    @computed
-    get activeSourceCategories(): DocCategory[] {
-        if (!this.activeDoc) return [];
-        const {source} = this.activeDoc;
-        return this.docService
-            .getCategories(source)
-            .filter(cat => this.docService.getDocsByCategory(source, cat.id).length > 0);
-    }
-
-    /** H2-level sections parsed from the current document content. */
-    @computed
-    get sections(): DocSection[] {
-        if (!this.content) return [];
-        return extractSections(this.content);
+    /** Sync grid selection + exit search whenever a doc is activated. */
+    protected override onDocActivated(entry: DocEntry) {
+        const recId = `${entry.source}:${entry.id}`;
+        this.gridModel.selectAsync(recId, {ensureVisible: true});
+        this.searchMode = false;
     }
 
     /** Toolbox example tabs relevant to the active doc. */
     @computed
     get activeDocExamples(): DocExampleLink[] {
-        if (!this.activeDoc) return [];
-        return getDocExamples(this.activeDoc.id);
-    }
-
-    /**
-     * Navigate to a specific doc by source + ID, optionally scrolling to a section (an H2 slug).
-     * Updates grid selection, content, and route. If the requested doc is already active, only
-     * the section scroll is (re)triggered.
-     */
-    @action
-    navigateToDoc(docId: string, source?: string, section?: string) {
-        const entry = source
-            ? this.docService.getDocEntry(docId, source)
-            : this.docService.getDocEntry(docId);
-        if (!entry) return;
-
-        const sameDoc = entry.id === this.activeDoc?.id && entry.source === this.activeDoc?.source;
-        if (sameDoc) {
-            if (section) this.pendingScrollSection = section;
-            return;
-        }
-
-        const recId = `${entry.source}:${entry.id}`;
-        this.gridModel.selectAsync(recId, {ensureVisible: true});
-
-        this.activeDoc = entry;
-        this.activeSection = null;
-        this.pendingScrollSection = section ?? null;
-        this.searchMode = false;
-        this.loadContentAsync(entry);
-        this.updateRouteFromDoc();
-    }
-
-    /** Clear the pending scroll-to-section request, once the view has consumed it. */
-    @action
-    clearPendingScrollSection() {
-        this.pendingScrollSection = null;
-    }
-
-    /** Navigate to the first doc in a given source + category. */
-    navigateToCategory(source: string, categoryId: string) {
-        const firstDoc = this.docService.getDocsByCategory(source, categoryId)[0];
-        if (firstDoc) this.navigateToDoc(firstDoc.id, firstDoc.source);
+        return this.activeDoc ? getDocExamples(this.activeDoc.id) : [];
     }
 
     /** Toggle search mode on/off. */
@@ -261,12 +154,6 @@ export class DocsPanelModel extends HoistModel {
                 this.exitSearchMode();
                 break;
         }
-    }
-
-    /** Update the active section - called from scroll-based observation in the view. */
-    @action
-    setActiveSection(sectionId: string) {
-        this.activeSection = sectionId;
     }
 
     /**
@@ -453,62 +340,6 @@ export class DocsPanelModel extends HoistModel {
         this.navigateToDoc(docId, source);
     }
 
-    /** Route → doc: sync active doc when the URL changes. */
-    private updateDocFromRoute() {
-        const {name, params} = XH.routerState;
-        if (!name.startsWith(this.BASE_ROUTE)) return;
-
-        const ref = this.docRefFromRoute(params);
-        if (ref) this.navigateToDoc(ref.docId, ref.source, ref.section);
-    }
-
-    /** Doc → route: push active doc ID into the URL. */
-    private updateRouteFromDoc() {
-        const {activeDoc, BASE_ROUTE} = this,
-            {name} = XH.routerState;
-
-        if (!name.startsWith(BASE_ROUTE)) return;
-
-        if (activeDoc) {
-            XH.navigate(
-                `${BASE_ROUTE}.docRef`,
-                {source: activeDoc.source, docId: this.docIdToRoute(activeDoc.id)},
-                {replace: true}
-            );
-        } else {
-            XH.navigate(BASE_ROUTE, {replace: true});
-        }
-    }
-
-    private docIdToRoute(docId: string): string {
-        return docId.replaceAll('/', '~');
-    }
-
-    private docRefFromRoute(params: Record<string, string>): {
-        source: string;
-        docId: string;
-        section: string;
-    } {
-        const {source, docId, section} = params;
-        if (!source || !docId) return null;
-        return {source, docId: docId.replaceAll('~', '/'), section: section ?? null};
-    }
-
-    private async loadContentAsync(entry: DocEntry) {
-        runInAction(() => (this.content = null));
-        try {
-            const content = await this.docService
-                .fetchContentAsync(entry.source, entry.id)
-                .linkTo(this.loadContentTask);
-            runInAction(() => (this.content = content));
-        } catch (e) {
-            runInAction(() => {
-                this.content = `## Error Loading Document\n\nFailed to load **${entry.title}**.\n\n\`${e.message}\``;
-            });
-            XH.handleException(e, {showAlert: false});
-        }
-    }
-
     getCategoryIcon(categoryId: string) {
         switch (categoryId) {
             case 'app-development':
@@ -556,45 +387,4 @@ export class DocsPanelModel extends HoistModel {
                 return Icon.folder();
         }
     }
-}
-
-//------------------
-// Section parsing
-//------------------
-/**
- * Parse H2 headings from raw markdown content into navigable sections.
- */
-function extractSections(content: string): DocSection[] {
-    const regex = /^## (.+)$/gm;
-    const sections: DocSection[] = [],
-        slugCounts = new Map<string, number>();
-    let match: RegExpExecArray;
-    while ((match = regex.exec(content)) !== null) {
-        const title = stripInlineMarkdown(match[1].trim());
-        let id = slugify(title);
-        const count = slugCounts.get(id) || 0;
-        if (count > 0) id += `-${count}`;
-        slugCounts.set(id, count + 1);
-        sections.push({id, title});
-    }
-    return sections;
-}
-
-/** Remove inline markdown formatting (code, bold, italic, links) to get plain text. */
-function stripInlineMarkdown(text: string): string {
-    return text
-        .replace(/`([^`]*)`/g, '$1')
-        .replace(/\*\*([^*]*)\*\*/g, '$1')
-        .replace(/\*([^*]*)\*/g, '$1')
-        .replace(/\[([^\]]*)\]\([^)]*\)/g, '$1')
-        .trim();
-}
-
-/** Convert text to a URL-safe slug: lowercase, strip special chars, hyphenate spaces. */
-function slugify(text: string): string {
-    return text
-        .toLowerCase()
-        .replace(/[^\w\s-]/g, '')
-        .replace(/[\s-]+/g, '-')
-        .replace(/^-+|-+$/g, '');
 }
